requirements.txt
h5py==2.10.0
matplotlib==3.1.1
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.25.3
Pillow==6.1.0
scikit-image==0.16.1
scikit-learn==0.21.3
scipy==1.1.0
tensorboardX==1.9
toolz==0.10.0
torch==1.2.0
torchvision==0.2.1
